requirements.txt
streamlit
feedparser
google-genai